Refactored code.
Added package for config.
