+++ apache-http-client/src/test/java/io/refactoring/http5/client/example/classic/BaseClassicExampleTests.java
@@ -1,7 +1,7 @@
 package io.refactoring.http5.client.example.classic;
 
 import io.refactoring.http5.client.example.BaseExampleTests;
-import io.refactoring.http5.client.example.classic.util.JsonUtils;
+import io.refactoring.http5.client.example.util.JsonUtils;
 
 abstract class BaseClassicExampleTests extends BaseExampleTests {
   protected final JsonUtils jsonUtils = new JsonUtils();
